ci(lint): satisfy golangci-lint v1.64.8
Run #3 surfaced 5 real lint issues after pinning to v1.64.8:

- 11 files needed gofmt (mostly stray tabs in comments + multi-line
  godoc indent — auto-applied)
- bundle.go:582 unconvert: int64(v) where v is already int64
- tokenizer.go isUnicodeSpace was unused (single-line wrapper around
  unicode.IsSpace; removed with its unicode import)

Net behaviour unchanged; all 17 packages still green.

+++ bundle/bundle.go
@@ -579,7 +579,7 @@ func readAny(cfg *config.Config, path string) any {
 			return true
 		}
 		if v := cfg.GetInt(path); v != 0 {
-			return int64(v)
+			return v
 		}
 		if v := cfg.GetFloat(path); v != 0 {
 			return v
